[fix] astype affects performance (#63)

+++ attention.py
@@ -65,4 +65,4 @@ def bias(causal_mask: mx.array, dtype: mx.Dtype) -> mx.array:
     Returns:
         attn_bias (batch_size, .., query_len, key_value_len)
     """
-    return mx.logical_not(causal_mask) * LOG_0
+    return (mx.logical_not(causal_mask) * LOG_0).astype(dtype)
